pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "game-genius-bookmarks"
version = "0.1.0"
description = "Bookmark management system for game analysis"
requires-python = ">=3.8"
dependencies = [
    "obsws-python",
    "Pillow",
    "python-dotenv",
]

[tool.pylint.messages_control]
disable = [
    "C0114",  # missing-module-docstring
    "C0115",  # missing-class-docstring
    "C0116",  # missing-function-docstring
    "C0103",  # invalid-name
    "C0301",  # line-too-long
    "C0415",  # top-level-import
    "C0303",  # trailing-whitespace
    "C0330",  # bad-continuation
    "C0326",  # bad-whitespace
    "W0611",  # unused-import
    "W0612",  # unused-variable
    "W0613",  # unused-argument
    "W0621",  # redefined-outer-name
    "W0622",  # redefined-builtin
    "W0703",  # broad-except
    "W0702",  # bare-except
    "R0903",  # too-few-public-methods
    "R0913",  # too-many-arguments
    "R0914",  # too-many-locals
    "R0915",  # too-many-statements
    "R0912",  # too-many-branches
    "R0911",  # too-many-return-statements
    "R0902",  # too-many-instance-attributes
    "R0904",  # too-many-public-methods
    "R0801",  # duplicate-code
    "R0401",  # cyclic-import
    "R0905",  # too-many-ancestors
    "R0906",  # too-many-public-methods
    "R0907",  # too-many-public-methods
]

[tool.pyright]
# Disable type checking for this project
typeCheckingMode = "off"
# Or if you want to keep some type checking but ignore specific errors:
# reportUnknownArgumentType = false
# reportUnknownVariableType = false
# reportUnknownMemberType = false
# reportMissingTypeStubs = false

[tool.ruff]
select = ["E", "F", "I"]
# E: pycodestyle errors (mostly about code style, e.g., indentation, whitespace, line length, etc.)
# F: pyflakes errors (about code correctness, e.g., undefined names, unused imports, etc.)
# I:isort rules (for import sorting and grouping)

[tool.ruff.lint.pycodestyle]
# max-line-length = 140
max-line-length = 700

# [tool.ruff.lint.pyflakes]

# [tool.isort]
# include_trailing_comma = false
